requirements.txt
acme==0.26.0
asn1crypto==0.22.0
certbot==0.26.1
certbot-nginx==0.25.0
certifi==2019.6.16
chardet==3.0.4
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==1.9
funcsigs==0.4
future==0.15.2
idna==2.5
josepy==1.0.1
language-selector==0.1
mock==1.3.0
parsedatetime==2.4
pbr==1.8.0
pycurl==7.43.0
PyExecJS==1.5.1
pygobject==3.20.0
PyICU==1.9.2
pyOpenSSL==17.3.0
pyparsing==2.0.3
pyRFC3339==1.0
python-apt==1.1.0b1+ubuntu0.16.4.1
pytz==2014.10
requests==2.20.0
requests-toolbelt==0.8.0
six==1.10.0
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.3
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
